+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `Csv.LineReader` and `Csv.LineWriter` to clarify API
+- Add simpler factories to `Csv.Reader` and `Csv.Writer`
 
 ### Changed